feat(typing): convert test file js to ts
