+++ jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-};
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    testRegex: '.*\\.test\\.ts$',
+}
