+++ test/.eslintrc.yml
@@ -1,7 +1,8 @@
 root: true
+parser: '@typescript-eslint/parser'
 parserOptions:
-    ecmaVersion: 2018
-    sourceType: script
+    project: './tsconfig.json'
+    sourceType: module
 env:
     node: true
     es6: true
